requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
numpy==1.26.1
Pillow==10.1.0
PyMuPDF==1.23.6
python-dotenv==1.0.0
gunicorn==21.2.0
python-jose[cryptography]
sentry-sdk